Move cookies helpers to network utils
