+++ src/bidiMapper/domains/network/NetworkRequest.ts
@@ -34,12 +34,12 @@ import type {Result} from '../../../utils/result.js';
 import type {CdpTarget} from '../context/CdpTarget.js';
 import type {EventManager} from '../session/EventManager.js';
 
-import {Cookie} from './Cookie.js';
 import type {NetworkStorage} from './NetworkStorage.js';
 import {
   computeHeadersSize,
   bidiNetworkHeadersFromCdpFetchHeaders,
   bidiNetworkHeadersFromCdpNetworkHeaders,
+  cdpToBiDiCookie,
 } from './NetworkUtils.js';
 
 /** Abstracts one individual network request. */
@@ -690,6 +690,6 @@ export class NetworkRequest {
       .filter(({blockedReasons}) => {
         return !Array.isArray(blockedReasons) || blockedReasons.length === 0;
       })
-      .map(({cookie}) => Cookie.cdpToBiDiCookie(cookie));
+      .map(({cookie}) => cdpToBiDiCookie(cookie));
   }
 }
